Add basic index page and jquery. Update RuleBook
Decided RuleBook check methods should be passed a grid rather than
a RuleBookook initialises with a grid.

After adding a basic html page and jquery and playing the game there are
a few bugs. Too much logic is in the Game model and as a result, the
the interface works, but has errors. The next step is to move some of
the methods to the interface instead of having them called within
 the game's functions.